+++ gepetuto/main.py
@@ -22,10 +22,18 @@ def parse_args(args=None) -> argparse.Namespace:
         description="Tutorial edition framework",
     )
     parser.add_argument(
+        "-q",
+        "--quiet",
+        action="count",
+        default=int(os.environ.get("QUIET", 0)),
+        help="decrement verbosity level",
+    )
+
+    parser.add_argument(
         "-v",
         "--verbose",
         action="count",
-        default=0,
+        default=int(os.environ.get("VERBOSITY", 0)),
         help="increment verbosity level",
     )
     parser.add_argument(
@@ -85,11 +93,7 @@ def parse_args(args=None) -> argparse.Namespace:
 
     args = parser.parse_args(args=args)
 
-    if args.verbose == 0:
-        level = os.environ.get("GEPETUTO_LOG_LEVEL", "WARNING")
-    else:
-        level = 30 - 10 * args.verbose
-    logging.basicConfig(level=level)
+    logging.basicConfig(level=30 - 10 * args.verbose + 10 * args.quiet)
 
     LOG.debug("parsed arguments: %s", args)